requirements.txt
ruamel.yaml
easydict
scipy
pandas
numpy
click
rich
pyfiglet
matplotlib
wandb
pytest
spacy